pyproject.toml
[project]
name = "fix_video"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click>=8.3.1",
    "pillow>=12.1.0",
    "pyinstaller>=6.17.0",
    "taskipy>=1.14.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
]

[project.scripts]
may = "main:main"


#######################################################
# LINTERS - RUFF | Isort | Black
#######################################################
[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I"]
# select = ["A", "B", "C", "D", "E", "F", "I", "N", "PL", "RUF", "S", "T", "W"]
lint.fixable = ["ALL"]
lint.ignore = ["D203", "D213", "T201", "PLE1205"]
lint.per-file-ignores = {}
# Same as Black.

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
exclude = [
    "tests/*",
    ".venv/*",
    "build",
    "dist",
    "node_modules"
]

# [tool.ruff.lint.pydocstyle]
# convertion="google"


[tool.black]
line-length = 120
extend-exclude='''
.temp-*
| venv
| .venv
'''

#######################################################
# Poetry Scripts & Taskipy
#######################################################
[tool.poetry.scripts]
syncro_git = "src.fix:syncro_git"

[tool.taskipy.tasks]
pre_lint = { cmd = "echo 'Black' && black .", help = "apply the black linter"}
lint = { cmd = "ruff check . --fix", help = "apply ruff linter"}
tree = "uv tree > .pip.tree"
gen = "task gen_fix_video && task gen_fix_image"
gen_fix_video = "uv run pyinstaller --onefile --clean --name fix_video.exe --target-arch x86_64 src/launcher.py"
gen_fix_image = "uv run pyinstaller --onefile --clean --name fix_image.exe --target-arch x86_64 src/launcher_fix_image.py"
gg = 'uv run --frozen pyinstaller --onefile --clean --name fix_video --add-data "src;src" --hidden-import=main --collect-submodules=click --collect-all=click src/main.py --target-arch x86_64'
